+++ lib/ui/polls/PollsHomePanel.dart
@@ -50,18 +50,18 @@ class _PollsHomePanelState extends State<PollsHomePanel> implements Notification
   _PollType? _selectedPollType;
 
   List<Poll>? _myPolls;
-  PollsCursor? _myPollsCursor;
+  String? _myPollsCursor;
   String? _myPollsError;
   bool _myPollsLoading = false;
   
   List<Poll>? _recentPolls;
   List<Poll>? _recentLocalPolls;
-  PollsCursor? _recentPollsCursor;
+  String? _recentPollsCursor;
   String? _recentPollsError;
   bool _recentPollsLoading = false;
 
   List<Poll>? _groupPolls;
-  PollsCursor? _groupPollsCursor;
+  String? _groupPollsCursor;
   String? _groupPollsError;
   bool _groupPollsLoading = false;
   List<Group>? _myGroups;
@@ -597,7 +597,7 @@ class _PollsHomePanelState extends State<PollsHomePanel> implements Notification
     if (((_groupPolls == null) || (_groupPollsCursor != null)) && !_groupPollsLoading) {
       _setGroupPollsLoading(true);
       _loadMyGroupsIfNeeded().then((_) {
-        Set<String>? groupIds = _myGroupIds;
+        List<String>? groupIds = _myGroupIds;
         if (CollectionUtils.isNotEmpty(groupIds)) {
           Polls().getGroupPolls(groupIds, cursor: _groupPollsCursor)!.then((PollsChunk? result) {
             if (result != null) {
@@ -631,10 +631,10 @@ class _PollsHomePanelState extends State<PollsHomePanel> implements Notification
     _myGroups = await Groups().loadGroups(myGroups: true);
   }
 
-  Set<String>? get _myGroupIds {
-    Set<String>? groupIds;
+  List<String>? get _myGroupIds {
+    List<String>? groupIds;
     if (CollectionUtils.isNotEmpty(_myGroups)) {
-      groupIds = <String>{};
+      groupIds = [];
       _myGroups!.forEach((group) {
         groupIds!.add(group.id!);
       });